Choose destinations, updated Footer addon.
When clicking "Copy to Destination Folder", added the ability to specify which destination folders you want to copy your saved files to. It will then only copy the files that are destined to the destination folders you selected.

Slightly updated The builtin footer addon to support custom footer text from within the app.